+++ ChangeLog.md
@@ -4,6 +4,7 @@
 * Better handling of null values when adding elements to dataset
 * Fix default values when accessing SS/US element values (Anders Gustafsson, Cureos AB)
 * Fix decoding of JPEG2000 images with signed pixel data (Mahesh Dubey)
+* Ability to decompress single frame from dataset (Mahesh Dubey)
 
 #### v1.0.24 (10/01/2012)
 * Change the default presentation context transfer syntax acceptance behavior to prefer the SCU proposed order